+++ VNUA.FITA.MQTT.HRM/VNUA.FITA.MQTT.HRM/Controllers/TaiKhoanController.cs
@@ -23,17 +23,9 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
         }
         public bool KiemTranChucNang(int? idChucNang)
         {
-
-            string tk= HttpContext.Session.GetString("SessionUser");
-            var count = _context.NhanViens.Count(m=>m.TenTaiKhoan==tk&m.PhanQuyen==idChucNang);
-
-            ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
-            ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
-            ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
-            var user = HttpContext.Session.GetString("SessionUser");
-            var counts = _context.NhanViens.Count(m =>m.TenTaiKhoan==user&m.PhanQuyen == idChucNang);
-
-            if (counts == 0)
+            string tk = HttpContext.Session.GetString("SessionUser");
+            var count = _context.NhanViens.Count(m => m.TenTaiKhoan == tk & m.PhanQuyen == idChucNang);
+            if (count == 0)
             {
                 return false;
             }
@@ -45,16 +37,14 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
         // GET: TaiKhoan
         public async Task<IActionResult> Index(string sortOder, string searchString, string currentFilter, int? pageNumber)
         {
-
-            
-            if (KiemTranChucNang(1)==false)
-            {
-                return RedirectToAction("BaoLoi","BaoLoi");
-            }
-
             ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
+
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(2) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
 
             ViewData["CurrentSort"] = sortOder;
             if (searchString != null)
@@ -88,6 +78,11 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
 
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(2) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -112,6 +107,11 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
 
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(2) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
+
             ViewData["IdBP"] = new SelectList(_context.BoPhans, "IdBoPhan", "IdBoPhan");
             return View();
         }
@@ -129,6 +129,11 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
             ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
+
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(2) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
 
             if (ModelState.IsValid)
             {
@@ -148,6 +153,12 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
             ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
+
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(2) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -172,6 +183,12 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
             ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
+
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(2) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
+
             if (id != nhanVien.IdNhanVien)
             {
                 return NotFound();
@@ -207,6 +224,12 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
             ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
+
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(2) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -231,6 +254,12 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
             ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
+
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(2) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
+
             var nhanVien = await _context.NhanViens.FindAsync(id);
             _context.NhanViens.Remove(nhanVien);
             await _context.SaveChangesAsync();
@@ -242,6 +271,9 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
             ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
+
+            
+
             return _context.NhanViens.Any(e => e.IdNhanVien == id);
         }
